// src/app/components/home/home.enums.ts
export enum TrafficDensity {
  High = 'High',
  Medium = 'Medium',
  Low = 'Low'
}

export enum TrafficLocation {
  North = 'North',
  South = 'South',
  Northwest = 'Northwest',
  Northeast = 'Northeast',
  Southwest = 'Southwest',
  Southeast = 'Southeast'
}

export enum VehicleType {
  Car = 'Cars',
  Bike = 'Bikes'
}

export enum VehicleFilter {
  ALL, CAR, BIKE
}